requirements.txt
PyYAML==5.4.1
pytest==6.2.4